requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
dependency-injector==4.42.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
multidict==6.1.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
TgCrypto==1.2.5
typing_extensions==4.12.2
yarl==1.13.1
